requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Caching==2.1.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
statsmodels==0.14.0
openpyxl==3.1.2
xlsxwriter==3.1.2
